// src/app/modules/dashboard/mantenimiento/models/acceso.ts
export class Acceso{
    codigo : number = 0;
    dni : string = "";
    nombres : string = "";
    codigoCargo: string = "";
    cargo : string = "";
    codigoSede: string = "";
    sede: string = "";
    codigoSenamhi: string = "";
    usuario: string = "";
    contrasenia : string = "";
    fechaCreacion : string = "";
    estado : boolean = true;
    rol: string = "";
}